// frontend/src/services/OpenApi/models/OpeningDetailsStockingLayerDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CodeDescriptionDto } from './CodeDescriptionDto';
export type OpeningDetailsStockingLayerDto = {
    layer: CodeDescriptionDto;
    minWellspacedTrees: number | null;
    minPreferredWellspacedTrees: number | null;
    minHorizontalDistanceWellspacedTrees: number | null;
    targetWellspacedTrees: number | null;
    minResidualBasalArea: number | null;
    minPostspacingDensity: number | null;
    maxPostspacingDensity: number | null;
    maxConiferous: number | null;
    heightRelativeToComp: number | null;
};